+++ neoforge/src/main/java/net/sefacestudios/somemoreblocks/mixin/RotatedPillarBlockMixin.java
@@ -1,7 +1,6 @@
 package net.sefacestudios.somemoreblocks.mixin;
 
 import net.minecraft.world.item.context.UseOnContext;
-import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.RotatedPillarBlock;
 import net.minecraft.world.level.block.state.BlockState;
 import net.neoforged.neoforge.common.ItemAbility;
@@ -9,16 +8,15 @@ import net.neoforged.neoforge.common.extensions.IBlockExtension;
 import net.sefacestudios.somemoreblocks.block.RotatedCarvedPaleOakBlock;
 import net.sefacestudios.somemoreblocks.registries.SMBRegistries;
 import org.jetbrains.annotations.Nullable;
+import org.spongepowered.asm.mixin.Implements;
+import org.spongepowered.asm.mixin.Interface;
 import org.spongepowered.asm.mixin.Mixin;
 
 @Mixin(RotatedPillarBlock.class)
-public abstract class RotatedPillarBlockMixin extends Block implements IBlockExtension {
-  public RotatedPillarBlockMixin(Properties properties) {
-    super(properties);
-  }
+@Implements(@Interface(iface = IBlockExtension.class, prefix = "smb$neo$"))
+public abstract class RotatedPillarBlockMixin {
 
-  @Override
-  public @Nullable BlockState getToolModifiedState(BlockState state, UseOnContext ctx, ItemAbility itemAbility, boolean simulate) {
+  public @Nullable BlockState smb$neo$getToolModifiedState(BlockState state, UseOnContext ctx, ItemAbility itemAbility, boolean simulate) {
     return SMBRegistries.CARVED_BLOCKS.getNext(state.getBlock())
       .map((block) -> {
         if (block instanceof RotatedCarvedPaleOakBlock) {
@@ -26,6 +24,6 @@ public abstract class RotatedPillarBlockMixin extends Block implements IBlockExt
             .setValue(RotatedCarvedPaleOakBlock.MOON_PHASE, RotatedCarvedPaleOakBlock.currentMoonPhase(ctx.getLevel()));
         }
         return block.withPropertiesOf(state);
-      }).orElse(super.getToolModifiedState(state, ctx, itemAbility, simulate));
+      }).orElse(null);
   }
 }