+++ neoforge/src/main/java/net/sefacestudios/somemoreblocks/mixin/TinyCactusBlockMixin.java
@@ -8,13 +8,15 @@ import net.minecraft.world.level.pathfinder.PathType;
 import net.neoforged.neoforge.common.extensions.IBlockExtension;
 import net.sefacestudios.somemoreblocks.block.TinyCactusBlock;
 import org.jetbrains.annotations.Nullable;
+import org.spongepowered.asm.mixin.Implements;
+import org.spongepowered.asm.mixin.Interface;
 import org.spongepowered.asm.mixin.Mixin;
 
 @Mixin(TinyCactusBlock.class)
-public abstract class TinyCactusBlockMixin implements IBlockExtension {
+@Implements(@Interface(iface = IBlockExtension.class, prefix = "smb$neo$"))
+public abstract class TinyCactusBlockMixin {
 
-  @Override
-  public @Nullable PathType getBlockPathType(BlockState state, BlockGetter level, BlockPos pos, @Nullable Mob mob) {
+  public @Nullable PathType smb$neo$getBlockPathType(BlockState state, BlockGetter level, BlockPos pos, @Nullable Mob mob) {
     return PathType.DAMAGING;
   }
 }